requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196
transformers==4.35.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0